# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Collab Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  displayName?: string;
  grade?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  admin: "admin",
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  displayName?: string;
  grade?: string;
  role: UserRole;
  bio?: string;
  avatarUrl?: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  planning: "planning",
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  category: string;
  status: ProjectStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  ownerId: number;
  ownerName?: string;
  memberCount: number;
  createdAt: string;
  updatedAt: string;
  tags?: string[];
  isMember?: boolean;
}

export type ProjectDetail = Project & {
  members?: User[];
};

export interface CreateProjectRequest {
  title: string;
  description: string;
  category: string;
  tags?: string[];
}

export type UpdateProjectRequestStatus =
  (typeof UpdateProjectRequestStatus)[keyof typeof UpdateProjectRequestStatus];

export const UpdateProjectRequestStatus = {
  planning: "planning",
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface UpdateProjectRequest {
  title?: string;
  description?: string;
  category?: string;
  status?: UpdateProjectRequestStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress?: number;
  tags?: string[];
}

export interface Comment {
  id: number;
  content: string;
  projectId: number;
  userId: number;
  username?: string;
  displayName?: string;
  createdAt: string;
}

export interface CreateCommentRequest {
  content: string;
}

export interface Experiment {
  id: number;
  title: string;
  description: string;
  results?: string;
  projectId: number;
  userId: number;
  username?: string;
  displayName?: string;
  fileUrl?: string;
  createdAt: string;
}

export interface CreateExperimentRequest {
  title: string;
  description: string;
  results?: string;
  fileUrl?: string;
}

export interface Message {
  id: number;
  content: string;
  projectId: number;
  userId: number;
  username?: string;
  displayName?: string;
  createdAt: string;
}

export interface SendMessageRequest {
  content: string;
}

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType];

export const ResourceType = {
  article: "article",
  video: "video",
  dataset: "dataset",
  tool: "tool",
  course: "course",
} as const;

export interface Resource {
  id: number;
  title: string;
  description: string;
  url: string;
  category: string;
  type: ResourceType;
  imageUrl?: string;
}