# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Collab Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  email: string;
  displayName?: string;
  grade?: string;
  role: UserRole;
  bio?: string;
  avatarUrl?: string;
  createdAt: Date;
}